Refactor code for parsing command line arguments. Add init_command_line_parser event that can be used by test scripts and plugins to add command line arguments.
